requirements.txt
openai_whisper==20231117
pytube==15.0.0
tiktoken==0.5.2
transformers==4.30.2
